requirements.txt
streamlit>=1.20
requests>=2.28
python-dotenv>=1.0